added view form for person
