+++ src/components/PersonForm/index.ts
@@ -1,2 +1,1 @@
-export {Navigation} from './Navigation';
 export {PersonForm} from './PersonForm';
