+++ src/hooks/index.ts
@@ -1,3 +1,3 @@
 export { useGoldCupApi } from "./useGoldCupApi";
-export type { Team, Player, UserDetails } from "./useGoldCupApi";
+export type { Team, Player, UserDetails, Person } from "./useGoldCupApi";
